% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARA_functions.R
\name{ZhaoNew}
\alias{ZhaoNew}
\title{Title}
\usage{
ZhaoNew(
  PredictiveCov,
  PrognosticCov,
  mu,
  theta,
  gamma,
  m0 = 10,
  p = 0.8,
  omega,
  isResponseBinary,
  isTreatmentNumber
)
}
\arguments{
\item{PredictiveCov}{a vector. It specifies patients' predictive covariate \code{X} and must be discretized}

\item{PrognosticCov}{a matrix. It specifies patients' \code{k} prognostic covariates code{Z} and must be discretized}

\item{mu}{a vector of length 2. It specifies the predetermined model coefficients of treatment effect \code{T} and \code{1-T} to generate response.}

\item{theta}{a vector of length 2. It specifies the predetermined model coefficients of \code{X} and \code{X*T} to generate response.}

\item{gamma}{a vector of length \code{k}. It specifies the predetermined model coefficients of \code{Z} and \code{X*T} to generate response.}

\item{m0}{an integer. If specifies the initial number of \code{2m0} patients that are assigned to treatments 1 and 2 with restrict randomization.}

\item{p}{an integer between 0.75 and 0.95. It is the probability of Efron's biased coin to assign patients to treatments 1 and 2.}

\item{omega}{a vector of length \code{k+2} that sums to 1. It specifies the weights placed on overall, on \code{k} covariate margin and within a stratum cell, respectively.}

\item{isResponseBinary}{a logical evaluating to TRUE or FALSE indicating whether the response is binary.}

\item{isTreatmentNumber}{a logical evaluating to TRUE or FALSE indicating whether the Treatment is denoted as 1 and 0.}
}
\value{
\item{data}{The patients' data after assignment.}
}
\description{
Title
}
\examples{
ZhaoNew(PredictiveCov=sample(c(1, -1), 1000, replace = T),PrognosticCov=replicate(3, sample(c(1, -1), 1000, replace = T)),
 beta=c(0.5,0.5,1,-0.5, 0.5, 0.5,0.5), isResponseBinary = T, isTreatmentNumber=T, omega=rep(0.2,5))
}
